refactor: 도메인별로 패키지 구조를 변경하라
