--- src/main/java/nextstep/subway/application/SectionService.java
@@ -1,30 +1,0 @@
-package nextstep.subway.application;
-
-import nextstep.subway.domain.Line;
-import nextstep.subway.domain.Section;
-import nextstep.subway.domain.Station;
-import nextstep.subway.dto.SectionRequest;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@Transactional
-public class SectionService {
-
-    private final StationService stationService;
-    private final LineService lineService;
-
-    public SectionService(StationService stationService, LineService lineService) {
-        this.stationService = stationService;
-        this.lineService = lineService;
-    }
-
-    public void addSection(Long lineId, SectionRequest sectionRequest) {
-        Station upStation = stationService.findById(sectionRequest.getUpStationId());
-        Station downStation = stationService.findById(sectionRequest.getDownStationId());
-        Section section = Section.of(upStation, downStation, sectionRequest.getDistance());
-
-        Line line = lineService.findOne(lineId);
-        line.addSection(section);
-    }
-}
