--- src/main/java/nextstep/subway/dto/LineResponse.java
@@ -1,57 +1,0 @@
-package nextstep.subway.dto;
-
-import java.util.ArrayList;
-import java.util.List;
-import nextstep.subway.domain.Line;
-
-public class LineResponse {
-
-    private Long id;
-    private String name;
-    private String color;
-    private List<StationResponse> stations;
-    private List<SectionResponse> sections;
-
-    public LineResponse(
-        Long id,
-        String name,
-        String color,
-        List<StationResponse> stations,
-        List<SectionResponse> sections
-    ) {
-        this.id = id;
-        this.name = name;
-        this.color = color;
-        this.stations = stations;
-        this.sections = sections;
-    }
-
-    public static LineResponse from(Line line) {
-        return new LineResponse(
-            line.getId(),
-            line.getName(),
-            line.getColor(),
-            new ArrayList<>(line.getStationResponses()),
-            line.getSectionResponses());
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public List<StationResponse> getStations() {
-        return stations;
-    }
-
-    public List<SectionResponse> getSections() {
-        return sections;
-    }
-}
