--- src/main/java/nextstep/subway/dto/SectionRequest.java
@@ -1,26 +1,0 @@
-package nextstep.subway.dto;
-
-public class SectionRequest {
-
-    private final Long upStationId;
-    private final Long downStationId;
-    private final Long distance;
-
-    public SectionRequest(Long upStationId, Long downStationId, Long distance) {
-        this.upStationId = upStationId;
-        this.downStationId = downStationId;
-        this.distance = distance;
-    }
-
-    public Long getUpStationId() {
-        return upStationId;
-    }
-
-    public Long getDownStationId() {
-        return downStationId;
-    }
-
-    public Long getDistance() {
-        return distance;
-    }
-}
