--- src/main/java/nextstep/subway/exception/CannotRegisterException.java
@@ -1,14 +1,0 @@
-package nextstep.subway.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class CannotRegisterException extends SubwayCustomException {
-
-    public CannotRegisterException(String message) {
-        super(HttpStatus.BAD_REQUEST, message);
-    }
-
-    public CannotRegisterException(ExceptionType exceptionType) {
-        super(HttpStatus.BAD_REQUEST, exceptionType);
-    }
-}
