+++ src/main/java/nextstep/subway/line/application/LineService.java
@@ -1,0 +1,73 @@
+package nextstep.subway.line.application;
+
+import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.stream.Collectors;
+import nextstep.subway.station.application.StationService;
+import nextstep.subway.line.domain.Line;
+import nextstep.subway.line.domain.LineRepository;
+import nextstep.subway.station.domain.Station;
+import nextstep.subway.line.dto.LineRequest;
+import nextstep.subway.line.dto.LineResponse;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+public class LineService {
+
+    private final LineRepository lineRepository;
+    private final StationService stationService;
+
+    public LineService(LineRepository lineRepository, StationService stationService) {
+        this.lineRepository = lineRepository;
+        this.stationService = stationService;
+    }
+
+    public LineResponse saveLine(LineRequest lineRequest) {
+        Station upStation = findStationById(lineRequest.getUpStationId());
+        Station downStation = findStationById(lineRequest.getDownStationId());
+        Line line = lineRepository.save(Line.of(lineRequest, upStation, downStation));
+
+        return LineResponse.from(line);
+    }
+
+    @Transactional(readOnly = true)
+    public List<LineResponse> findAllLines() {
+        List<Line> lines = lineRepository.findLineAndStations();
+        return lines.stream()
+            .map(LineResponse::from)
+            .collect(Collectors.toList());
+    }
+
+    @Transactional(readOnly = true)
+    public LineResponse findLine(Long id) {
+        Line line = findLineById(id);
+        return LineResponse.from(line);
+    }
+
+    @Transactional(readOnly = true)
+    public Line findOne(Long id) {
+        return lineRepository.findLineAndStationsById(id)
+            .orElseThrow(() -> new NoSuchElementException("지하철 노선을 찾을 수 없습니다."));
+    }
+
+    public void updateLine(Long id, LineRequest lineRequest) {
+        Line line = findLineById(id);
+        line.update(lineRequest.getName(), lineRequest.getColor());
+    }
+
+    public void deleteLine(Long id) {
+        Line line = findLineById(id);
+        lineRepository.delete(line);
+    }
+
+    private Line findLineById(Long id) {
+        return lineRepository.findById(id)
+            .orElseThrow(() -> new NoSuchElementException("지하철 노선을 찾을 수 없습니다."));
+    }
+
+    private Station findStationById(Long id) {
+        return stationService.findById(id);
+    }
+}
