+++ src/main/java/nextstep/subway/section/application/SectionService.java
@@ -1,0 +1,32 @@
+package nextstep.subway.section.application;
+
+import nextstep.subway.station.application.StationService;
+import nextstep.subway.line.application.LineService;
+import nextstep.subway.line.domain.Line;
+import nextstep.subway.section.domain.Section;
+import nextstep.subway.station.domain.Station;
+import nextstep.subway.section.dto.SectionRequest;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+public class SectionService {
+
+    private final StationService stationService;
+    private final LineService lineService;
+
+    public SectionService(StationService stationService, LineService lineService) {
+        this.stationService = stationService;
+        this.lineService = lineService;
+    }
+
+    public void addSection(Long lineId, SectionRequest sectionRequest) {
+        Station upStation = stationService.findById(sectionRequest.getUpStationId());
+        Station downStation = stationService.findById(sectionRequest.getDownStationId());
+        Section section = Section.of(upStation, downStation, sectionRequest.getDistance());
+
+        Line line = lineService.findOne(lineId);
+        line.addSection(section);
+    }
+}
