+++ src/main/java/nextstep/subway/section/dto/SectionRequest.java
@@ -1,0 +1,26 @@
+package nextstep.subway.section.dto;
+
+public class SectionRequest {
+
+    private final Long upStationId;
+    private final Long downStationId;
+    private final Long distance;
+
+    public SectionRequest(Long upStationId, Long downStationId, Long distance) {
+        this.upStationId = upStationId;
+        this.downStationId = downStationId;
+        this.distance = distance;
+    }
+
+    public Long getUpStationId() {
+        return upStationId;
+    }
+
+    public Long getDownStationId() {
+        return downStationId;
+    }
+
+    public Long getDistance() {
+        return distance;
+    }
+}
