+++ src/main/java/nextstep/subway/section/ui/SectionController.java
@@ -1,0 +1,28 @@
+package nextstep.subway.section.ui;
+
+import nextstep.subway.section.application.SectionService;
+import nextstep.subway.section.dto.SectionRequest;
+import nextstep.subway.section.dto.SectionResponse;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class SectionController {
+
+    private final SectionService sectionService;
+
+    public SectionController(SectionService sectionService) {
+        this.sectionService = sectionService;
+    }
+
+    @PostMapping("/lines/{lineId}/sections")
+    public ResponseEntity<SectionResponse> addSection(@PathVariable Long lineId,
+        @RequestBody SectionRequest sectionRequest) {
+        sectionService.addSection(lineId, sectionRequest);
+        return new ResponseEntity<>(HttpStatus.CREATED);
+    }
+}
