+++ src/main/java/nextstep/subway/station/domain/Station.java
@@ -1,0 +1,46 @@
+package nextstep.subway.station.domain;
+
+import java.util.Objects;
+import javax.persistence.*;
+import nextstep.subway.global.domain.BaseEntity;
+
+@Entity
+public class Station extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @Column(unique = true)
+    private String name;
+
+    protected Station() {
+    }
+
+    public Station(String name) {
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        Station station = (Station) o;
+        return Objects.equals(getId(), station.getId());
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(getId());
+    }
+}
