+++ src/main/java/nextstep/subway/station/dto/StationResponse.java
@@ -1,0 +1,47 @@
+package nextstep.subway.station.dto;
+
+import java.util.Objects;
+import nextstep.subway.station.domain.Station;
+
+public class StationResponse {
+
+    private Long id;
+    private String name;
+
+    public static StationResponse of(Station station) {
+        return new StationResponse(station.getId(), station.getName());
+    }
+
+    public StationResponse() {
+    }
+
+    public StationResponse(Long id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        StationResponse that = (StationResponse) o;
+        return Objects.equals(getId(), that.getId());
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(getId());
+    }
+}
