+++ src/test/java/nextstep/subway/section/SectionAcceptanceTest.java
@@ -19,8 +19,8 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.stream.Stream;
 import nextstep.subway.AcceptanceTest;
-import nextstep.subway.dto.LineResponse;
-import nextstep.subway.exception.ExceptionType;
+import nextstep.subway.line.dto.LineResponse;
+import nextstep.subway.global.exception.ExceptionType;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
